Refactoring of names
Instead of UseCase suffix, I changed the name in Service so to be more compliant with a DDD naming convention.
While I'm still unhappy with the Generator name that is not so much OO but since FizzBuzz is an artificial use case it is difficult to map it on a sound model.